% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_pmrm.R
\name{analyze_ps_pmrm}
\alias{analyze_ps_pmrm}
\title{Analyze trial data with the proportion slowing PMRM}
\usage{
analyze_ps_pmrm(data_trial, b_start = 0.2)
}
\arguments{
\item{data_trial}{A data frame from a single clinical trial. The following
variables should be present:
\itemize{
\item \code{ADAScog_integer}: integer-valued ADAScog score
\item \code{arm_time}: factor that characterizes the treatment arm-time combination.
\item \code{time_int}: integer that indicates the measurement occasion. Should start
with \code{1} for the first occasion, values for subsequent occasions should be
consecutive integers.
}}

\item{b_start}{Starting value for the proportional slowing parameter.
Defaults to 0.}
}
\value{
object returned by \code{nlme::gnls}
}
\description{
Analyze trial data with the proportion slowing PMRM
}
