% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_trials.R
\name{analyze_trials}
\alias{analyze_trials}
\title{Analyze a list of trials}
\usage{
analyze_trials(list_data_trials, trial_function, n_cores = 1, ...)
}
\arguments{
\item{list_data_trials}{A list of data frames}

\item{trial_function}{A function that takes the data frame as argument, fits
a model, and returns the corresponding fitted-model object.}

\item{n_cores}{The number of cores used for parallel computing. If
\code{n_cores = 1}, no parallel computng is performed.}

\item{...}{Additional parameters that are passed on to \code{trial_function()}.}
}
\value{
A list of fitted-model objects.
}
\description{
The \code{\link[=analyze_trials]{analyze_trials()}} function analyzes a list of data frames where each
data frame corresponds to a single clinical trial. The fitted-model object
for each trial is returned as a list of model-fit objects. That list can
further be used to extract the relevant estimates or p-values for each trial.
}
