% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_utils.R
\name{BC_percentile_CI}
\alias{BC_percentile_CI}
\title{Bias-corrected percentile confidence intervals}
\usage{
BC_percentile_CI(estimate, boot_replicates, alpha = 0.05)
}
\arguments{
\item{estimate}{The point estimate of the target parameter.}

\item{boot_replicates}{A numeric vector with bootstrap replicates of the
estimate of interest.}

\item{alpha}{Significance level for the confidence interval. Default is \code{0.05}.}
}
\value{
(numeric) vector with lower and upper limits
}
\description{
The \code{\link[=BC_percentile_CI]{BC_percentile_CI()}} function computes the bias-corrected percentile
bootstrap confidence interval given the bootstrap replicates.
}
