% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_utils.R
\name{bootstrap_CI}
\alias{bootstrap_CI}
\title{Bootstrap confidence intervals}
\usage{
bootstrap_CI(boot_replicates, estimate, alpha = 0.05, type = "BCa")
}
\arguments{
\item{boot_replicates}{A numeric vector with bootstrap replicates of the
estimate of interest.}

\item{estimate}{The point estimate of the target parameter.}

\item{alpha}{Significance level for the confidence interval. Default is \code{0.05}.}

\item{type}{Type of bootstrap confidence interval to compute. Currently only
two types are implemented: \code{"BC percentile"} and \code{"percentile"}.}
}
\value{
(numeric) vector with lower and upper limits
}
\description{
The \code{\link[=bootstrap_CI]{bootstrap_CI()}} function computes bootstrap confidence intervals given
the bootstrap replicates.
}
