% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_vertical_estimator.R
\name{constrained_vertical_estimator}
\alias{constrained_vertical_estimator}
\title{Constrained vertical Estimator}
\usage{
constrained_vertical_estimator(alpha_obs, beta_obs, Sigma_obs)
}
\description{
Constrained vertical Estimator
}
