% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_vertical_estimator.R
\name{constraint_matrix_A1}
\alias{constraint_matrix_A1}
\title{Return constraint matrix for Assumption 1}
\usage{
constraint_matrix_A1(length_alpha, length_beta)
}
\arguments{
\item{length_beta}{}
}
\description{
Return constraint matrix for Assumption 1
}
