% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_vertical_estimator.R
\name{constraint_matrix_A2}
\alias{constraint_matrix_A2}
\title{Return Constraint matrix for Assumption 2}
\usage{
constraint_matrix_A2(length_alpha, length_beta)
}
\arguments{
\item{length_beta}{}
}
\description{
Return Constraint matrix for Assumption 2
}
