% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast-test.R
\name{contrast_estimate_common_se}
\alias{contrast_estimate_common_se}
\title{Standard Error of contrast-based estimator of the common acceleration factor}
\usage{
contrast_estimate_common_se(
  gamma_est,
  time_points,
  ctrl_estimates,
  exp_estimates,
  interpolation,
  vcov,
  type = "omnibus",
  j = 1:length(exp_estimates),
  weights = NULL,
  ...
)
}
\arguments{
\item{gamma_est}{Estimated value for the common acceleration factor.}

\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_estimates} vector. The first element
should be zero and correspond to the time of randomization.}

\item{ctrl_estimates}{Estimated mean outcome in the control group at the
fixed occasions corresponding to the times in \code{time_points}.}

\item{exp_estimates}{Estimated mean outcomes in the experimental group at
fixed occasions corresponding to the times in \code{time_points[-1]}. Note that
first element in \code{exp_estimates} should correspond to some time after
randomization.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation.
\item \code{"spline"}: natural cubic spline interpolation. This interpolation method has been most
thoroughly tested is most stable.
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson.
}}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}. If an element of \code{c(ctrl_means, exp_means)}
is known (e.g., mean at baseline is zero when using change from baseline as
outcome), then the corresponding row and column in \code{vcov} should be set to
zero.}

\item{type}{Which type of test statistic should be used. See Test Statistic
Variants. Should be one of
\enumerate{
\item \code{type = "omnibus"}
\item \code{type = "directional"}
\item \code{type = "inverse variance"}
\item \code{type = "custom"}
}}

\item{j}{(Integer) vector that indicates which elements in \code{exp_estimates}
should be used for the contrast test. Defaults to \code{1:length(exp_estimates)},
i.e., all elements are used.}

\item{weights}{If \code{type == "custom"}, the user should specify a weight
vector for weighting estimates at different time points differently.}

\item{...}{Tuning parameters that are passed to \code{\link[stats:optim]{stats::optim()}}.}
}
\value{
(numeric) Estimated SE of the estimator.
}
\description{
Standard Error of contrast-based estimator of the common acceleration factor
}
