% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast-test.R
\name{contrast_test_common}
\alias{contrast_test_common}
\title{contrast test for common acceleration factor}
\usage{
contrast_test_common(
  time_points,
  ctrl_estimates,
  exp_estimates,
  ref_fun,
  interpolation,
  vcov,
  gamma_0 = 1,
  type = "omnibus",
  j = 1:length(exp_estimates),
  weights = NULL
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_estimates} vector. The first element
should be zero and correspond to the time of randomization.}

\item{ctrl_estimates}{Estimated mean outcome in the control group at the
fixed occasions corresponding to the times in \code{time_points}.}

\item{exp_estimates}{Estimated mean outcomes in the experimental group at
fixed occasions corresponding to the times in \code{time_points[-1]}. Note that
first element in \code{exp_estimates} should correspond to some time after
randomization.}

\item{ref_fun}{Intra- and extrapolation function that is returned by
\code{ref_fun_constructor()}.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation.
\item \code{"spline"}: natural cubic spline interpolation. This interpolation method has been most
thoroughly tested is most stable.
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson.
}}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}. If an element of \code{c(ctrl_means, exp_means)}
is known (e.g., mean at baseline is zero when using change from baseline as
outcome), then the corresponding row and column in \code{vcov} should be set to
zero.}

\item{gamma_0}{Value for the acceleration factor under the null hypothesis.}

\item{type}{Which type of test statistic should be used. See Test Statistic
Variants. Should be one of
\enumerate{
\item \code{type = "omnibus"}
\item \code{type = "custom"}
}}

\item{j}{(Integer) vector that indicates which elements in \code{exp_estimates}
should be used for the contrast test. Defaults to \code{1:length(exp_estimates)},
i.e., all elements are used.}

\item{weights}{If \code{type == "custom"}, the user should specify a weight
vector for weighting estimates at different time points differently.}
}
\value{
Named (numeric) vector with two elements:
\enumerate{
\item \code{"z"} or \code{"chi-squared"}. The test statistic.
\item \code{"p-value"}: two-sided p-value.
}
}
\description{
The \code{\link[=contrast_test_common]{contrast_test_common()}} function implements the contrast test under the
assumption that there exists a common acceleration factor. Multiple variants
to this test exist and are implemented.
}
\section{Test Statistic Variants}{
In this documentation file, we continue with the distributional results
for the contrast functions as detailed in the documentation of
\code{\link[=contrast_test]{contrast_test()}}. The \code{\link[=contrast_test_common]{contrast_test_common()}} function implements
several types of tests for the null hypothesis that the common acceleration
factor equals a particular value, \eqn{\gamma_0}. Note that in the functions
available to users, only the \code{type = "custom"} test is available (other tests
have not been thoroughly studied and are, therefore, not direct available).

Two types of test statistics are implemented in the \code{\link[=contrast_test_common]{contrast_test_common()}}
function:
\enumerate{
\item \code{type = "omnibus"}
\item \code{type = "custom"}
}

These are discussed in more detail next.
\subsection{Omnibus}{

The omnibus contrast test is based on the "classic" chi-squared statistic that is
defined as follows:
\deqn{
n
\boldsymbol{\Phi}(\gamma_0; \hat{\boldsymbol{\alpha}}_n, \hat{\boldsymbol{\beta}}_n)^\top
\Sigma_n^{-1}
\boldsymbol{\Phi}(\gamma_0; \hat{\boldsymbol{\alpha}}_n, \hat{\boldsymbol{\beta}}_n).
}
which follows asymptotically a chi-squared distribution with K degrees of
freedom if \eqn{\gamma_0} is the true common acceleration factor. An
important feature of this test
statistic is that it reduces to the chi-squared statistic for \eqn{H_0 :
\alpha_j = \beta_j} for all \eqn{j} in \code{j} when \code{gamma_0 = 1}. However, this
also means that there is no gain in power.
}

\subsection{Custom}{

The "custom" contrast test allows for user-specified weights, \eqn{w}. The test
statistic is defined as follows:
\deqn{
n^{1/2}
\left(\boldsymbol{w}^\top \Omega_n \boldsymbol{w}\right)^{-1/2})
\boldsymbol{w}^\top
\boldsymbol{\Phi}(\gamma_0; \hat{\boldsymbol{\alpha}}_n, \hat{\boldsymbol{\beta}}_n),
}
which follows asymptotically a standard normal distribution if
\eqn{\gamma_0} is the true common acceleration factor.
}
}

