% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{delta_vertical_to_horizontal}
\alias{delta_vertical_to_horizontal}
\title{Transform vertical treatment effect estimates to time scale}
\usage{
delta_vertical_to_horizontal(
  time_points,
  ctrl_means,
  exp_means,
  vcov,
  interpolation = "spline"
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_means} vector.}

\item{ctrl_means}{Estimated mean outcome in the control group at fixed
occasions.}

\item{exp_means}{Estimated mean outcomes in the experimental group at fixed
occasions.}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation
\item \code{"spline"}: natural cubic spline interpolation
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson
}}
}
\value{
A list with three element:
\itemize{
\item \code{estimate}: the transformed parameter estimates
\item \code{variance}: the variance-covariance matrix for the transformed parameter
estimates
\item \code{partial}: the Jacobian matrix
}
}
\description{
This function transform the vertical parameter estimates to parameter
estimates on the time scale.
}
