% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extrapol_fun_factory}
\alias{extrapol_fun_factory}
\title{Function factory for the extrapolating part of the reference profile}
\usage{
extrapol_fun_factory(x_ref, y_ref)
}
\arguments{
\item{x_ref}{Vector of reference time values.}

\item{y_ref}{Vector of reference (mean) outcome values.}
}
\value{
A (linear) R function.
}
\description{
This functions returns a function that extrapolates beyond the range of the
observed time points in \code{.x_ref}. This function is the linear function
through the first and last point of the reference profile. This ensures some
stability in the extrapolation which is not present when some interpolation
methods are used for extrapolation.
}
