% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_4PL}
\alias{fit_4PL}
\title{Fit 4 PL function to observed reference profile.}
\usage{
fit_4PL(x_ref, y_ref)
}
\arguments{
\item{x_ref}{Vector of reference time values.}

\item{y_ref}{Vector of reference (mean) outcome values.}
}
\value{
A vector containing the four estimates for the 4PL function
parameters.
}
\description{
The \code{\link[=fit_4PL]{fit_4PL()}} function fits the four parameter logistic function to the
data points in \code{x_ref} and \code{y_ref}. This is done by minimizing the sum of
deviations between the 4PL function and the values in \code{y_ref}.
}
