% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{four_pl_f_factory}
\alias{four_pl_f_factory}
\title{Function factory for 4 PL function}
\usage{
four_pl_f_factory(x_ref, y_ref)
}
\arguments{
\item{x_ref}{Vector of reference time values.}

\item{y_ref}{Vector of reference (mean) outcome values.}
}
\value{
A 4 PL function.
}
\description{
Function factory for 4 PL function
}
