% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{g_Delta_bis}
\alias{g_Delta_bis}
\title{Compute acceleration factor}
\usage{
g_Delta_bis(par, method, time_points)
}
\arguments{
\item{par}{Vector with estimated points of the trajectories at \code{time_points}.
This vector is a combinations of the estimates in both treatment groups.
The \code{length(time_points)} first elements correspond to the reference group.
The \code{length(time_points) - 1} remaining elements correspond to the active
treatment group.}

\item{method}{Interpolation method; see \code{\link[=TCT]{TCT()}}.}

\item{time_points}{Time points to which the first \code{length(time_points)}
elements in \code{par} correspond.}
}
\value{
A vector of length \code{length(time_points) - 1} containing the estimated
acceleration factors for the estimates in the active treatment group.
}
\description{
The \code{\link[=g_Delta_bis]{g_Delta_bis()}} function computes the time-specific acceleration factor
from the estimated trajectories in both treatment groups.
}
