% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_new_time}
\alias{get_new_time}
\title{Maps observed outcomes to a reference profile using interpolation}
\usage{
get_new_time(y_ref, x_ref, y_obs, method = "linear")
}
\arguments{
\item{y_ref}{Vector of reference (mean) outcome values.}

\item{x_ref}{Vector of reference time values.}

\item{y_obs}{Vector of (mean) outcome values to be mapped to the reference
profile.}

\item{method}{Method for interpolation:
\itemize{
\item \code{"linear"}: piecewise linear interpolation
\item \code{"spline"}: natural cubic spline interpolation
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch and
Carlson.
\item \verb{"fourPL}: four parameter logistic regression. This can only be used when
the reference profile contains four or more means, i.e.,
\code{length(y_ref) >= 4}. When there are more than four means, this is
strictly speaking no longer interpolation. The corresponding parameters
are found by minimizing the sum of squared deviations from the four PL
curve.
}}
}
\value{
A vector of the time-mapped values.
}
\description{
The \code{\link[=get_new_time]{get_new_time()}} function maps observed (mean) outcomes to the time-axis
of a reference profile. The reference profile is defined by a set of
time-outcome pairs in the reference group. These pairs are usually the
estimated mean outcomes at fixed time-points in the control group.
}
\section{Reference Profile}{
Let \eqn{f_{0}(t)} denote the reference profile. This
profile is estimated as the interpolating curve between the points specified
in \code{x_ref} and \code{y_ref}. The interpolation method is specified in the \code{method}
argument.
}

\section{Time Mapping}{
Let \eqn{y^{\ast}} denote the observed (mean) outcome. The time-mapping is
then defined as \eqn{t^{\ast} = f^{-1}_{0}(y^{\ast})}. This is equivalent to
solving \eqn{f_{0}(t^{\ast}) = y^{\ast}} for \eqn{t^{\ast}}. The latter
approach is implemented numerically with \code{\link[stats:uniroot]{stats::uniroot()}} for all
\code{method}s, expect for \code{"fourPL"}. For the latter method,
\eqn{f^{-1}_{0}(\cdot)} is obtained analytically.
}

