% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-test.R
\name{gradient_g}
\alias{gradient_g}
\title{Compute Gradient for Score Test}
\usage{
gradient_g(t_m, x_ref, y_ref, method = "spline")
}
\description{
The \code{\link[=gradient_g]{gradient_g()}} function computes
\eqn{\nabla g_{\gamma}(\boldsymbol{\alpha}, \boldsymbol{\beta}; t_j)^t} where
\eqn{g_{\gamma}(\boldsymbol{\alpha}, \boldsymbol{\beta}; t_j) = \beta_j - f_0( \gamma \cdot t_{j}; \boldsymbol{\alpha})}.
}
