% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpol_fun_factory}
\alias{interpol_fun_factory}
\title{Function factory for the interpolating part of the reference profile}
\usage{
interpol_fun_factory(x_ref, y_ref, method)
}
\arguments{
\item{x_ref}{Vector of reference time values.}

\item{y_ref}{Vector of reference (mean) outcome values.}

\item{method}{Method for interpolation:
\itemize{
\item \code{"linear"}: piecewise linear interpolation
\item \code{"spline"}: natural cubic spline interpolation
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch and
Carlson.
\item \verb{"fourPL}: four parameter logistic regression. This can only be used when
the reference profile contains four or more means, i.e.,
\code{length(y_ref) >= 4}. When there are more than four means, this is
strictly speaking no longer interpolation. The corresponding parameters
are found by minimizing the sum of squared deviations from the four PL
curve.
}}
}
\value{
An interpolating R function.
}
\description{
This function returns a function that interpolates within the range of the
observed time points in \code{x_ref}. This is done using the interpolation method
in the \code{method} argument.
}
