% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linear_interpolation_f_factory}
\alias{linear_interpolation_f_factory}
\title{Function factory for linear interpolation function}
\usage{
linear_interpolation_f_factory(x_ref, y_ref)
}
\arguments{
\item{x_ref}{Vector of reference time values.}

\item{y_ref}{Vector of reference (mean) outcome values.}
}
\value{
A linear interpolation R function.
}
\description{
Function factory for linear interpolation function
}
