% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear-gls.R
\name{nonlinear_gls_estimator}
\alias{nonlinear_gls_estimator}
\title{Common Acceleration Factor Estimation via Generalized Least Squares}
\usage{
nonlinear_gls_estimator(
  time_points,
  ctrl_estimates,
  exp_estimates,
  interpolation,
  vcov,
  j = 1:length(exp_estimates),
  gamma_0 = NULL,
  start_gamma = 0.75,
  ...
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_estimates} vector. The first element
should be zero and correspond to the time of randomization.}

\item{ctrl_estimates}{Estimated mean outcome in the control group at the
fixed occasions corresponding to the times in \code{time_points}.}

\item{exp_estimates}{Estimated mean outcomes in the experimental group at
fixed occasions corresponding to the times in \code{time_points[-1]}. Note that
first element in \code{exp_estimates} should correspond to some time after
randomization.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation.
\item \code{"spline"}: natural cubic spline interpolation. This interpolation method has been most
thoroughly tested is most stable.
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson.
}}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}. If an element of \code{c(ctrl_means, exp_means)}
is known (e.g., mean at baseline is zero when using change from baseline as
outcome), then the corresponding row and column in \code{vcov} should be set to
zero.}

\item{j}{(Integer) vector that indicates which elements in \code{exp_estimates}
should be used for the contrast test. Defaults to \code{1:length(exp_estimates)},
i.e., all elements are used.}

\item{gamma_0}{Defaults to \code{NULL}. If a non-null value is given to this argument,
the common acceleration factor is held constant at that value.}

\item{start_gamma}{Starting value for the common acceleration factor in the
optimization procedure. Defaults to \code{0.75}.}

\item{...}{Tuning parameters that are passed to \code{\link[stats:optim]{stats::optim()}}.}
}
\value{
A list with the following elements:
\itemize{
\item \code{estimates}: Estimated parameters by minimizing the generalized least
squares criterion. The first \code{length(ctrl_estimates)} values are the estimated
means in the control group. The last element is the estimated common acceleration
factor.
\item \code{criterion}: Value of the criterion function in the parameter estimates.
}
}
\description{
\code{\link[=nonlinear_gls_estimator]{nonlinear_gls_estimator()}} estimates the common acceleration factor through
minimizing a generalized least-squares criterion.
}
\details{
In the following sections, we use notation introduced in the documentation of
\code{\link[=TCT_meta]{TCT_meta()}}.
}
\section{Generalized Least Squares}{
The generalized least squares (GLS) estimator \eqn{(\hat{\boldsymbol{\alpha}}_n, \hat{\gamma}_n)}
is defined as the minimizer of the following criterion:

\deqn{
(\hat{\boldsymbol{\alpha}}_n, \hat{\gamma}_n)
= \arg \min_{(\boldsymbol{\alpha}, \gamma)} \left\{
\begin{pmatrix}
\hat{\boldsymbol{\alpha}}_n \\
\hat{\boldsymbol{\beta}}_n
\end{pmatrix}
-
\begin{pmatrix}
\boldsymbol{\alpha} \\
\boldsymbol{f}_0(\gamma \cdot \boldsymbol{t}; \boldsymbol{\alpha})
\end{pmatrix}
\right\}^\top \hat{\Sigma}_n^{-1}
\left\{
\begin{pmatrix}
\hat{\boldsymbol{\alpha}}_n \\
\hat{\boldsymbol{\beta}}_n
\end{pmatrix}
-
\begin{pmatrix}
\boldsymbol{\alpha} \\
\boldsymbol{f}_0(\gamma \cdot \boldsymbol{t}; \boldsymbol{\alpha})
\end{pmatrix}
\right\},
}
where
\eqn{\boldsymbol{f}_0(\gamma \cdot \boldsymbol{t}; \boldsymbol{\alpha}) =
(f_0(\gamma \cdot t_1; \boldsymbol{\alpha}), \dots, f_0(\gamma \cdot t_K; \boldsymbol{\alpha}))'}

This criterion jointly estimates \eqn{\boldsymbol{\alpha}} and \eqn{\gamma},
even though one is generally only interested in the common acceleration
factor, \eqn{\gamma}.
}

\section{Inference}{
One can use the minimized GLS criterion for hypothesis testing and confidence
intervals. Let \eqn{Q_{1, n}} be the above GLS criterion minimized in
\eqn{\boldsymbol{\alpha}} and \eqn{\gamma} and multiplied with \eqn{n}, and
let \eqn{Q_{0, n}(\gamma_0)} be the corresponding GLS criterion minimized in
\eqn{\boldsymbol{\alpha}} while holding \eqn{\gamma} constant at
\eqn{\gamma_0}. Then, under the null hypothesis that
\eqn{\gamma = \gamma_0}, the test statistic \eqn{T_n = Q_{0, n}(\gamma_0) -
Q_{1, n}} follows asymptotically a chi-squared distribution with one degree
of freedom (under regularity conditions). This test statistic is also used
for computing confidence intervals for the common acceleration factor.

Along the same arguments as in the previous paragraph, we have that
\eqn{Q_{1, n}} follows a chi-squared distribution with \eqn{K - 1} degrees of
freedom under the proportional-slowing assumption (and a correct
interpolation method). This result is used for testing the
proportional-slowing assumption.
}

