% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear-gls.R
\name{nonlinear_gls_estimator_vcov}
\alias{nonlinear_gls_estimator_vcov}
\title{Variance-Covariance matrix of nonlinear GLS estimator}
\usage{
nonlinear_gls_estimator_vcov(
  time_points,
  interpolation,
  vcov,
  j,
  gamma_est,
  alpha_est
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_estimates} vector. The first element
should be zero which generally corresponds to the time of randomization.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation.
\item \code{"spline"}: natural cubic spline interpolation. This interpolation method has been most
thoroughly tested is most stable.
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson.
}}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}.}

\item{j}{(Integer) vector that indicates which elements in \code{exp_estimates}
should be used for the score test. Defaults to \code{1:length(exp_estimates)},
i.e., all elements are used.}

\item{gamma_est}{Estimated value for the common acceleration factor.}

\item{alpha_est}{Estimated value of the mean vector in the control group.
Note that this should be the estimate as returned by
\code{\link[=nonlinear_gls_estimator]{nonlinear_gls_estimator()}}, and not the original estimates.}
}
\value{
(numeric) variance-covariance matrix where the rows and columns
correspond to
\eqn{(\hat{\alpha}_0, \hat{\alpha}_1, ..., \hat{\alpha}_K, \hat{\gamma})'}.
}
\description{
Variance-Covariance matrix of nonlinear GLS estimator
}
