% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear-gls.R
\name{nonlinear_gls_test}
\alias{nonlinear_gls_test}
\title{Compute test statistic and p-value for the nonlinear GLS estimator}
\usage{
nonlinear_gls_test(
  time_points,
  ctrl_estimates,
  exp_estimates,
  interpolation,
  vcov,
  j = 1:length(exp_estimates),
  gamma_0 = 1,
  start_gamma = 0.75,
  ...
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_estimates} vector. The first element
should be zero which generally corresponds to the time of randomization.}

\item{ctrl_estimates}{Estimated mean outcome in the control group at the
fixed occasions corresponding to the times in \code{time_points}.}

\item{exp_estimates}{Estimated mean outcomes in the experimental group at
fixed occasions corresponding to the times in \code{time_points[-1]}. Note that
first element in \code{exp_estimates} should correspond to some time after
randomization.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation.
\item \code{"spline"}: natural cubic spline interpolation. This interpolation method has been most
thoroughly tested is most stable.
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson.
}}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}.}

\item{j}{Measurement occasion to test acceleration factor for. This
corresponds to \code{exp_estimates[j]}.}

\item{gamma_0}{Value for the acceleration factor under the null hypothesis.}

\item{start_gamma}{Starting value for the acceleration factor used for for the
generalized least squares estimator.}

\item{...}{Tuning parameters that are passed to \code{\link[stats:optim]{stats::optim()}}.}
}
\value{
Named (numeric) vector with two elements:
\enumerate{
\item \code{"z"} or \code{"chi-squared"}. The test statistic.
\item \code{"p-value"}: two-sided p-value.
}
}
\description{
Compute test statistic and p-value for the nonlinear GLS estimator
}
