% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_utils.R
\name{percentile_CI}
\alias{percentile_CI}
\title{Percentile confidence intervals}
\usage{
percentile_CI(boot_replicates, alpha = 0.05)
}
\arguments{
\item{boot_replicates}{A numeric vector with bootstrap replicates of the
estimate of interest.}

\item{alpha}{Significance level for the confidence interval. Default is \code{0.05}.}
}
\value{
(numeric) vector with lower and upper limits
}
\description{
The \code{\link[=percentile_CI]{percentile_CI()}} function computes the percentile bootstrap confidence
interval given the bootstrap replicates.
}
