% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{pm_bootstrap_vertical_to_horizontal}
\alias{pm_bootstrap_vertical_to_horizontal}
\title{Title}
\usage{
pm_bootstrap_vertical_to_horizontal(
  time_points,
  ctrl_estimates,
  exp_estimates,
  vcov,
  interpolation = "spline",
  B = 100
)
}
\arguments{
\item{B}{}
}
\description{
Title
}
