% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{print.summary_TCT_meta}
\alias{print.summary_TCT_meta}
\title{Print summary_TCT_meta object}
\usage{
\method{print}{summary_TCT_meta}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{\link[=summary.TCT_meta]{summary.TCT_meta()}}.}

\item{...}{
  Arguments passed on to \code{\link[base:print]{base::print}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Print summary_TCT_meta object
}
