% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional-slowing-test.R
\name{proportional_slowing_gls_test}
\alias{proportional_slowing_gls_test}
\title{Test for proportional slowing based on GLS criterion}
\usage{
proportional_slowing_gls_test(
  time_points,
  ctrl_estimates,
  exp_estimates,
  interpolation,
  vcov,
  j,
  start_gamma
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_estimates} vector. The first element
should be zero which generally corresponds to the time of randomization.}

\item{ctrl_estimates}{Estimated mean outcome in the control group at the
fixed occasions corresponding to the times in \code{time_points}.}

\item{exp_estimates}{Estimated mean outcomes in the experimental group at
fixed occasions corresponding to the times in \code{time_points[-1]}. Note that
first element in \code{exp_estimates} should correspond to some time after
randomization.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation.
\item \code{"spline"}: natural cubic spline interpolation. This interpolation method has been most
thoroughly tested is most stable.
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson.
}}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}.}

\item{j}{(Integer) vector that indicates which elements in \code{exp_estimates}
should be used for the score test. Defaults to \code{1:length(exp_estimates)},
i.e., all elements are used.}

\item{start_gamma}{Starting value for the acceleration factor used for for the
generalized least squares estimator.}
}
\value{
A named vector with the following elements:
\itemize{
\item \code{chi_squared}: The chi-squared test statistic that asymptotically follows
a chi-squared distribution with \code{df} degrees of freedom under proportional
slowing.
\item \code{df}: Degrees of freedom of the chi-squared distribution under proportional
slowing.
\item \code{p_value}: The p-value based on the asymptotic chi-squared distribution
with \code{df} degrees of freedom.
}
}
\description{
\code{\link[=proportional_slowing_gls_test]{proportional_slowing_gls_test()}} performs a test for proportional slowing
based on the minimized GLS criterion.
}
