% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-test.R
\name{score_estimate_common}
\alias{score_estimate_common}
\title{Estimate the common acceleration factor by minimizing the squared score
statistic}
\usage{
score_estimate_common(
  time_points,
  ctrl_estimates,
  exp_estimates,
  ref_fun,
  interpolation,
  vcov,
  type = "omnibus",
  j = 1:length(exp_estimates),
  weights = NULL,
  penalty = function(x) 0,
  ...
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_estimates} vector. The first element
should be zero which generally corresponds to the time of randomization.}

\item{ctrl_estimates}{Estimated mean outcome in the control group at the
fixed occasions corresponding to the times in \code{time_points}.}

\item{exp_estimates}{Estimated mean outcomes in the experimental group at
fixed occasions corresponding to the times in \code{time_points[-1]}. Note that
first element in \code{exp_estimates} should correspond to some time after
randomization.}

\item{ref_fun}{Intra- and extrapolation function that is returned by
\code{ref_fun_constructor()}.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation.
\item \code{"spline"}: natural cubic spline interpolation. This interpolation method has been most
thoroughly tested is most stable.
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson.
}}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}.}

\item{type}{Which type of test statistic should be used. See Test Statistic
Variants. Should be one of
\enumerate{
\item \code{type = "omnibus"}
\item \code{type = "directional"}
\item \code{type = "inverse variance"}
\item \code{type = "custom"}
}}

\item{j}{(Integer) vector that indicates which elements in \code{exp_estimates}
should be used for the score test. Defaults to \code{1:length(exp_estimates)},
i.e., all elements are used.}

\item{weights}{If \code{type == "custom"}, the user should specify a weight
vector for weighting estimates at different time points differently.}

\item{penalty}{This a function that is added to the (squared) test
statistics. Defaults to a constant function. This is mostly useful for
small samples to put a penalty on values of gamma outside the unit
interval.}

\item{...}{Tuning parameters that are passed to \code{\link[stats:optim]{stats::optim()}}.}
}
\value{
(numeric) estimated for the common acceleration factor.
}
\description{
The \code{\link[=score_estimate_common]{score_estimate_common()}} function estimates the common acceleration
factor. The estimate is the common acceleration factor which minimizes the
test statistic for the corresponding null hypothesis, i.e., the most likely
value of \eqn{\gamma} where "likelihood" is defined in terms of a test
statistic.
}
