% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-test.R
\name{score_estimate_common_se}
\alias{score_estimate_common_se}
\title{Standard Error of Score-based estimator of the common acceleration factor}
\usage{
score_estimate_common_se(
  gamma_est,
  time_points,
  ctrl_estimates,
  exp_estimates,
  interpolation,
  vcov,
  type = "omnibus",
  j = 1:length(exp_estimates),
  weights = NULL,
  ...
)
}
\arguments{
\item{gamma_est}{Estimated value for the common acceleration factor.}

\item{type}{Which type of test statistic should be used. See Test Statistic
Variants. Should be one of
\enumerate{
\item \code{type = "omnibus"}
\item \code{type = "directional"}
\item \code{type = "inverse variance"}
\item \code{type = "custom"}
}}

\item{j}{(Integer) vector that indicates which elements in \code{exp_estimates}
should be used for the score test. Defaults to \code{1:length(exp_estimates)},
i.e., all elements are used.}

\item{weights}{If \code{type == "custom"}, the user should specify a weight
vector for weighting estimates at different time points differently.}

\item{...}{Tuning parameters that are passed to \code{\link[=optim]{optim()}}.}
}
\value{
(numeric) Estimated SE of the estimator.
}
\description{
Standard Error of Score-based estimator of the common acceleration factor
}
