% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-test-utils.R
\name{score_vector_jacobian}
\alias{score_vector_jacobian}
\title{Compute Jacobian of the score vector}
\usage{
score_vector_jacobian(
  time_points,
  ctrl_estimates,
  exp_estimates,
  interpolation,
  vcov,
  gamma_0,
  j,
  weights
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_estimates} vector.}

\item{ctrl_estimates}{Estimated mean outcome in the control group at fixed
occasions.}

\item{exp_estimates}{Estimated mean outcomes in the experimental group at
fixed occasions.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation
\item \code{"spline"}: natural cubic spline interpolation
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson
}}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}.}

\item{gamma_0}{Value under the null hypothesis for the acceleration factor.}

\item{j}{description}

\item{weights}{description}
}
\description{
The \code{\link[=score_vector_jacobian]{score_vector_jacobian()}} function computes the Jacobian of the score
vector with respect to \eqn{(\boldsymbol{\alpha},\boldsymbol{\beta})'},
\deqn{\frac{\partial \boldsymbol{s}(\gamma \cdot \boldsymbol{t}; \boldsymbol{\alpha}, \boldsymbol{\beta})}{\partial (\boldsymbol{\alpha}, \boldsymbol{\beta})}}.
}
