% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_N_trials}
\alias{simulate_N_trials}
\title{Simulates N clinical trials under the supplied model parameters.}
\usage{
simulate_N_trials(
  control = "placebo",
  delta = 1.3,
  total_time = 50,
  n_measurements = 5,
  sample_size = 1000,
  N_trials = 1,
  details = FALSE
)
}
\arguments{
\item{control}{Type of control group, can be:
\itemize{
\item \code{"placebo"}: Control group receives placebo
\item \code{"symptomatic"}: control group receives a symptomatic treatment
}}

\item{delta}{deceleration factor. This parameterizes the disease-modifying
treatment effect.}

\item{total_time}{Total trial duration in weeks.}

\item{n_measurements}{Number of equally spaced measurement occasions for each
patient.}

\item{sample_size}{Total sample size in each trial.}

\item{N_trials}{Total number of trial to simulate independently.}

\item{details}{Return detailed data frame:
\itemize{
\item \code{TRUE}: the returned data frame contains more variables such as
the underlying random effect values and numeric values for sampled covariates
and outcome values.
\item \code{FALSE} (default): the returned data frame only contains variables that
are available in real clinical trials.
}}
}
\value{
Data frame with the same structure as \link{simulated_test_trial}.
}
\description{
Simulates N clinical trials under the supplied model parameters.
}
\examples{
# simulate a single trial with default arguments
simulate_N_trials()
}
