% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_single_trial}
\alias{simulate_single_trial}
\title{Simulates one clinical trial under the supplied model parameters.}
\usage{
simulate_single_trial(
  control = "placebo",
  delta = 1.3,
  total_time = 50,
  n_measurements = 5,
  sample_size = 1000,
  trial_number = 1,
  details = FALSE
)
}
\arguments{
\item{control}{Type of control group, can be:
\itemize{
\item \code{"placebo"}: Control group receives placebo
\item \code{"symptomatic"}: control group receives a symptomatic treatment
}}

\item{delta}{deceleration factor. This parameterizes the disease-modifying
treatment effect.}

\item{total_time}{Total trial duration in weeks.}

\item{n_measurements}{Number of equally spaced measurement occasions for each
patient.}

\item{sample_size}{Total sample size in each trial.}

\item{trial_number}{Defaults to 1. This is used by \code{\link[=simulate_N_trials]{simulate_N_trials()}} to
distinguish different trials, but is not relevant when
\code{\link[=simulate_single_trial]{simulate_single_trial()}} is used by the user directly to simulate a single
trial.}

\item{details}{Return detailed data frame:
\itemize{
\item \code{TRUE}: the returned data frame contains more variables such as
the underlying random effect values and numeric values for sampled covariates
and outcome values.
\item \code{FALSE} (default): the returned data frame only contains variables that
are available in real clinical trials.
}}
}
\value{
Data frame with the same structure as \link{simulated_test_trial}.
}
\description{
Simulates one clinical trial under the supplied model parameters.
}
\examples{
# simulate a single trial with default parameters
simulate_single_trial()

}
