% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_test_trial}
\alias{simulated_test_trial}
\title{Simulated clinical trial data}
\format{
\subsection{\code{simulated_test_trial} A data frame with 5,000 rows and 9 columns:}{

\describe{
\item{Week}{The number of weeks since randomization the measurement is taken.}
\item{SubjId}{Integer that uniquely identifies patients}
\item{Apo}{Number of ApoE4 alleles}
\item{Gender}{Gender of the patient}
\item{ADAScog_integer}{Measured score on the ADAScog scale}
\item{Age_integer}{Age at time of randomization}
\item{BMMSE}{Measured Mini Mental State Examination score at time of randomization}
\item{arm}{Integer that indicates treatment arm: 0 for control, 1 for active treatment}
\item{trial_number}{Integer that uniquely identifies trials. Not relevant since there is data for only a single trial.}
}
}
}
\usage{
simulated_test_trial
}
\description{
Data that was simulated with the \code{\link[=simulate_single_trial]{simulate_single_trial()}} function with the
following argument:
\itemize{
\item \code{control = "placebo"}
\item \code{delta = 1.3}
\item \code{total_time = 100}
\item \code{n_measurements = 5}
\item \code{sample_size = 1000}
\item \code{trial_number = 1}
\item \code{details = FALSE}
}
}
\details{
This is corresponds to a trial with 5 equally spaced measurements per patient
over a period of 100 weeks. The total sample size is 1000, and there are
patient groups: (i) placebo and (ii) disease-modifying treatment. The
treatment effect corresponds to a deceleration factor of 1.3.
}
\keyword{datasets}
