% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{summary.TCT_meta}
\alias{summary.TCT_meta}
\title{Summarize fitted Time Component Test model}
\usage{
\method{summary}{TCT_meta}(x, alpha = 0.05, delta_transformation = "identity")
}
\arguments{
\item{x}{Object returned by \code{\link[=TCT_meta]{TCT_meta()}}.}

\item{alpha}{Two-sided confidence level for confidence intervals.}

\item{delta_transformation}{Transformation when applying the delta-method to
obtain confidence intervals.}
}
\description{
Summarize fitted Time Component Test model
}
