% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{summary.TCT_meta_common}
\alias{summary.TCT_meta_common}
\title{Title}
\usage{
\method{summary}{TCT_meta_common}(x, alpha = 0.05, delta_transformation = "identity")
}
\arguments{
\item{alpha}{}
}
\description{
Title
}
