% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{TCT_common}
\alias{TCT_common}
\title{Estimate Common Acceleration Factor}
\usage{
TCT_common(
  TCT_Fit,
  B = 0,
  bs_fix_vcov = FALSE,
  select_coef = 1:length(coef(TCT_Fit)),
  null_bs = FALSE,
  gls_est = TRUE,
  constraints = FALSE
)
}
\arguments{
\item{bs_fix_vcov}{}
}
\description{
The \code{\link[=TCT_common]{TCT_common()}} function estimates the common acceleration factor.
}
