% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{TCT_meta_common}
\alias{TCT_meta_common}
\title{Estimate Common Acceleration Factor}
\usage{
TCT_meta_common(
  TCT_Fit,
  inference = "wald",
  B = 0,
  bs_fix_vcov = FALSE,
  select_coef = 1:length(coef(TCT_Fit)),
  gls_est = TRUE,
  constraints = FALSE,
  type = NULL,
  weights = NULL
)
}
\arguments{
\item{bs_fix_vcov}{}
}
\description{
The \code{\link[=TCT_meta_common]{TCT_meta_common()}} function estimates the common acceleration factor.
}
