% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{trajectory}
\alias{trajectory}
\title{Trajectory function}
\usage{
trajectory(t, z, gamma, alpha, x_ref, method)
}
\arguments{
\item{t}{Time at which the trajectory's value is computed.}

\item{z}{Group indicater:
\itemize{
\item \code{z = 0}: indicates the reference group
\item \code{z = 1}: indicates the treatment group
}}

\item{gamma}{Acceleration factor parameter.}

\item{alpha}{(Estimated) points on the reference profile.}

\item{x_ref}{Vector of reference time values.}

\item{method}{Method for interpolation:
\itemize{
\item \code{"linear"}: piecewise linear interpolation
\item \code{"spline"}: natural cubic spline interpolation
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch and
Carlson.
\item \verb{"fourPL}: four parameter logistic regression. This can only be used when
the reference profile contains four or more means, i.e.,
\code{length(y_ref) >= 4}. When there are more than four means, this is
strictly speaking no longer interpolation. The corresponding parameters
are found by minimizing the sum of squared deviations from the four PL
curve.
}}
}
\value{
(numeric) value of the trajctory at \code{t}.
}
\description{
This function returns the value of the trajectory at \code{t}.
}
