% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm.R
\name{analyze_mmrm}
\alias{analyze_mmrm}
\title{Analyze trial data with MMRM}
\usage{
analyze_mmrm(data_trial, method = "ML", type = "full")
}
\arguments{
\item{data_trial}{A data frame from a single clinical trial. The following
variables should be present:
\itemize{
\item \code{ADAScog_integer}: integer-valued ADAScog score
\item \code{arm_time}: factor that characterizes the treatment arm-time combination.
\item \code{time_int}: integer that indicates the measurement occasion. Should start
with \code{1} for the first occasion, values for subsequent occasions should be
consecutive integers.
}}

\item{method}{Which method should be used for fitting the MMRM? Should be one
of the following strings:
\itemize{
\item \code{"ML"} (default): maximum likelihood
\item \code{"REML"}: restricted maximum likelihood
}}

\item{type}{Which MMRM model should be use?
\itemize{
\item \code{"full"} (default): full model with treatment-time interaction.
\item \code{"null"}: null model with no treatment effect. The only use case for
fitting this model is to perform likelihood ratio tests.
}}
}
\value{
fitted-model object from \code{\link[nlme:gls]{nlme::gls()}}
}
\description{
The \code{\link[=analyze_mmrm]{analyze_mmrm()}} function analyzes data from a single trial with the
mixed model for repeated measures (See Details).
}
