% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast-test.R
\name{contrast_test}
\alias{contrast_test}
\title{Compute TCT contrast test z-value}
\usage{
contrast_test(
  time_points,
  ctrl_estimates,
  exp_estimates,
  ref_fun,
  interpolation,
  vcov,
  j,
  gamma_0 = 1
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_estimates} vector. The first element
should be zero and correspond to the time of randomization.}

\item{ctrl_estimates}{Estimated mean outcome in the control group at the
fixed occasions corresponding to the times in \code{time_points}.}

\item{exp_estimates}{Estimated mean outcomes in the experimental group at
fixed occasions corresponding to the times in \code{time_points[-1]}. Note that
first element in \code{exp_estimates} should correspond to some time after
randomization.}

\item{ref_fun}{Intra- and extrapolation function that is returned by
\code{ref_fun_constructor()}.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation.
\item \code{"spline"}: natural cubic spline interpolation. This interpolation method has been most
thoroughly tested is most stable.
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson.
}}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}. If an element of \code{c(ctrl_means, exp_means)}
is known (e.g., mean at baseline is zero when using change from baseline as
outcome), then the corresponding row and column in \code{vcov} should be set to
zero.}

\item{j}{Measurement occasion to test acceleration factor for. This
corresponds to \code{exp_estimates[j]}.}

\item{gamma_0}{Value for the acceleration factor under the null hypothesis.}
}
\value{
Named (numeric) vector with two elements:
\enumerate{
\item \code{"z"}. This test-statistic follows a standard normal distribution under
the null hypothesis.
\item \code{"p-value"}: two-sided p-value.
}
}
\description{
The \code{\link[=contrast_test]{contrast_test()}} function computes the z-value for the contrast test at
measurement \code{j}. This corresponds to \code{exp_estimates[j]}. For technical
details on the contrast test, see Details.
}
\details{
In the following sections, we use notation introduced in the documentation of
\code{\link[=TCT_meta]{TCT_meta()}}.
}
\section{Contrast Test}{
For constructing a contrast test, we start from the null hypothesis,
\deqn{H_{\gamma_0}: \gamma_j = \gamma_0,} where \eqn{\gamma_j} is the
acceleration factor at time \eqn{t_j}. Under this null hypothesis, and taking
the control group as reference, we can
compute the experimental treatment group mean at \eqn{t_j} as \deqn{\beta_{0,
j} = f_0( \gamma_0 \cdot t_{j}; \boldsymbol{\alpha})} given the true control
group mean vector, \eqn{\boldsymbol{\alpha}}. Consequently, a test for
\eqn{H_0: \beta_j = \beta_{0, j}} is also a valid test for
\eqn{H_{\gamma_0}}. Well-established tests are available for the former
hypothesis. However, these cannot be applied directly since
\eqn{\boldsymbol{\alpha}} is not known, but only estimated. Indeed, \eqn{\beta_{0,
j}} is itself estimated by \eqn{\hat{\beta}_{0, j} =  f_0( \gamma_0 \cdot t_{j};
\hat{\boldsymbol{\alpha}})}. Still, this provides the starting point for a
contrast test that draws upon well-established statistical tests.

We first define the contrast vector as follows, \deqn{\boldsymbol{\Phi}(\boldsymbol{\gamma}; \hat{\boldsymbol{\alpha}}_n, \hat{\boldsymbol{\beta}}_n) =
\hat{\boldsymbol{\beta}}_n - \left(f_0(\gamma_1 \cdot t_1;
\hat{\boldsymbol{\alpha}}_n), \dots, f_0(\gamma_K \cdot t_K;
\hat{\boldsymbol{\alpha}}_n)\right)^\top.}
Under \eqn{H_{\gamma_0}} for all time points and using the delta
method, we have that
\deqn{n^{1/2}\boldsymbol{\Phi}(\boldsymbol{\gamma_0}; \hat{\boldsymbol{\alpha}}_n,
\hat{\boldsymbol{\beta}}_n) \overset{d}{\to} \mathcal{N}(\boldsymbol{0}, \Omega)} where
\deqn{\Omega =
\dot{\boldsymbol{\Phi}}_{\boldsymbol{\alpha}, \boldsymbol{\beta}}(\boldsymbol{\gamma_0}; \boldsymbol{\alpha_0}, \boldsymbol{\beta_0}) \cdot \Sigma \cdot
\dot{\boldsymbol{\Phi}}_{\boldsymbol{\alpha}, \boldsymbol{\beta}}(\boldsymbol{\gamma_0}; \boldsymbol{\alpha_0}, \boldsymbol{\beta_0})^\top} with
\eqn{\dot{\boldsymbol{\Phi}}_{\boldsymbol{\alpha}, \boldsymbol{\beta}}(\boldsymbol{\gamma_0}; \boldsymbol{\alpha_0}, \boldsymbol{\beta_0})} the
Jacobian of \eqn{(\boldsymbol{\alpha}, \boldsymbol{\beta}) \mapsto \boldsymbol{\Phi}(\boldsymbol{\gamma_0}; \boldsymbol{\alpha}, \boldsymbol{\beta})} evaluated in the true parameter vector,
\eqn{(\boldsymbol{\alpha_0}, \boldsymbol{\beta_0})'}.
In practice, \eqn{\Omega} is unknown because the true parameter vector is unknown.
We, therefore, replace the unkown parameters wit the corresponding estimates in
the above expression for \eqn{\Omega}; the corresponding estimated matrix is
denoted by \eqn{\hat{\Omega}_n}. As long as \eqn{\hat{\Omega}_n} is consistent
for \eqn{\Omega}, we have the following distributional result, which forms the
basis for hypothesis tests and confidence intervals:
\deqn{n^{1/2}\boldsymbol{\Phi}(\boldsymbol{\gamma_0}; \hat{\boldsymbol{\alpha}}_n,
\hat{\boldsymbol{\beta}}_n) \overset{d}{\to} \mathcal{N}(\boldsymbol{0}, \hat{\Omega}_n).}
Specifically, for tests for time-specific acceleration factors, we use the
following distributional result:
\deqn{n^{1/2}\Phi_j(\gamma_0; \hat{\boldsymbol{\alpha}}_n,
\hat{\boldsymbol{\beta}}_n) \overset{d}{\to} \mathcal{N}(\boldsymbol{0}, \hat{\Omega}_n)}
where
}

