% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{DeltaMethod}
\alias{DeltaMethod}
\title{Delta method with numerical derivatives}
\usage{
DeltaMethod(par, fct, vcov, ...)
}
\arguments{
\item{par}{vector of parameter estimates}

\item{fct}{function that transforms the parameter values to the appropriate
function}

\item{vcov}{The variance-covariance matrix of the parameter estimates in the
\code{par} vector argument}

\item{...}{}
}
\value{
A list with three element:
\itemize{
\item \code{estimate}: the transformed parameter estimates
\item \code{variance}: the variance-covariance matrix for the transformed parameter
estimates
\item \code{partial}: the Jacobian matrix
}
}
\description{
Delta method with numerical derivatives
}
