% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-test-utils.R
\name{f0_gradient_t}
\alias{f0_gradient_t}
\title{Gradient of the reference trajectory at fixed time point}
\usage{
f0_gradient_t(ref_fun, t_vec)
}
\arguments{
\item{t_vec}{}
}
\description{
The \code{\link[=f0_gradient_t]{f0_gradient_t()}} function is a simple wrapper that returns
\eqn{\frac{\partial \boldsymbol{f_0} (\boldsymbol{t^*}; \boldsymbol{\alpha}, \boldsymbol{\beta})}{\partial \boldsymbol{t^*}}}.
}
