% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta-method.R
\name{g_Delta_bis}
\alias{g_Delta_bis}
\title{Compute acceleration factor}
\usage{
g_Delta_bis(par, interpolation, time_points)
}
\arguments{
\item{par}{Vector with estimated points of the trajectories at \code{time_points}.
This vector is a combinations of the estimates in both treatment groups.
The \code{length(time_points)} first elements correspond to the reference group.
The \code{length(time_points) - 1} remaining elements correspond to the active
treatment group.}

\item{time_points}{Time points to which the first \code{length(time_points)}
elements in \code{par} correspond.}
}
\value{
A (numeric) vector of length \code{length(time_points) - 1} that
corresponds to \eqn{\boldsymbol{\gamma}}.
}
\description{
The \code{\link[=g_Delta_bis]{g_Delta_bis()}} function computes the time-specific acceleration factor
from the estimated trajectories in both treatment groups.
}
\details{
We can write the slowing factor at \eqn{t_j} as a function of the the mean
parameters,
\deqn{
\boldsymbol{\gamma} = \boldsymbol{\gamma}(\boldsymbol{t}; \boldsymbol{\alpha}, \boldsymbol{\beta}).
}
The \code{\link[=g_Delta_bis]{g_Delta_bis()}} function computes \eqn{\boldsymbol{\gamma}} where \code{par}
corresponds to \eqn{(\boldsymbol{\alpha}, \boldsymbol{\beta})'} and
\code{time_points} corresponds to \eqn{\boldsymbol{t}}.
}
