% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{g_Delta_bis}
\alias{g_Delta_bis}
\title{Transform vertical treatment effect estimates to time scale}
\usage{
g_Delta_bis(par, method, time_points)
}
\description{
This function transform the vertical parameter estimates to parameter
estimates on the time scale.
}
