% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear-gls.R
\name{nonlinear_gls_estimator}
\alias{nonlinear_gls_estimator}
\title{TCT Estimation by Nonlinear GLS}
\usage{
nonlinear_gls_estimator(
  time_points,
  ctrl_estimates,
  exp_estimates,
  interpolation,
  vcov,
  j = 1:length(exp_estimates),
  gamma_0 = NULL,
  start_gamma = 0.75,
  ...
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_estimates} vector. The first element
should be zero which generally corresponds to the time of randomization.}

\item{ctrl_estimates}{Estimated mean outcome in the control group at the
fixed occasions corresponding to the times in \code{time_points}.}

\item{exp_estimates}{Estimated mean outcomes in the experimental group at
fixed occasions corresponding to the times in \code{time_points[-1]}. Note that
first element in \code{exp_estimates} should correspond to some time after
randomization.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation.
\item \code{"spline"}: natural cubic spline interpolation. This interpolation method has been most
thoroughly tested is most stable.
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson.
}}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}. If an element of \code{c(ctrl_means, exp_means)}
is known (e.g., mean at baseline is zero when using change from baseline as
outcome), then the corresponding row and column in \code{vcov} should be set to
zero.}

\item{j}{(Integer) vector that indicates which elements in \code{exp_estimates}
should be used for the score test. Defaults to \code{1:length(exp_estimates)},
i.e., all elements are used.}

\item{gamma_0}{Defaults to \code{NULL}. If a non-null value is given to this argument,
the common acceleration factor is held constant at that value.}

\item{start_gamma}{Starting value for the acceleration factor used for for the
generalized least squares estimator.}

\item{...}{Tuning parameters that are passed to \code{\link[stats:optim]{stats::optim()}}.}
}
\value{
A list with the following elements:
\itemize{
\item \code{estimates}: Estimated parameters by minimizing the generalized least
squares criterion. The first \code{length(ctrl_estimates)} values are the estimated
means in the control group. The last element is the estimated common acceleration
factor.
\item \code{criterion}: Value of the criterion function in the parameter estimates.
}
}
\description{
TCT Estimation by Nonlinear GLS
}
