% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_weights_score.R
\name{optimize_weights}
\alias{optimize_weights}
\title{Find the "optimal" weights for the score estimator}
\usage{
optimize_weights(
  time_points,
  ctrl_estimates,
  exp_estimates,
  ref_fun,
  interpolation,
  vcov,
  j = 1:length(exp_estimates),
  weights = rep(1, length(exp_estimates)),
  epsilon = 1e-06,
  ...
)
}
\arguments{
\item{ref_fun}{Intra- and extrapolation function that is returned by
\code{ref_fun_constructor()}.}

\item{j}{(Integer) vector that indicates which elements in \code{exp_estimates}
should be used for the score test. Defaults to \code{1:length(exp_estimates)},
i.e., all elements are used.}

\item{weights}{Staring values for the weights vector.}

\item{epsilon}{Tolerance to stop optimization algorithm. If the l2-norm of
the difference in weights from two subsequent iterations is smaller than
\code{epsilon}, the optimization algorithm stops.}
}
\value{
(numeric) vector with optimal weights.
}
\description{
The \code{\link[=optimize_weights]{optimize_weights()}} function finds the weights for the score estimator
that minimize the estimated variance of this estimator. The estimated
variance is based in the delta method.
}
