% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_weights_score.R
\name{optimize_weights}
\alias{optimize_weights}
\title{Find the "optimal" weights for the score estimator}
\usage{
optimize_weights(
  time_points,
  ctrl_estimates,
  exp_estimates,
  ref_fun,
  interpolation,
  vcov,
  j = 1:length(exp_estimates),
  weights = rep(1, length(exp_estimates)),
  epsilon = 1e-06
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_estimates} vector. The first element
should be zero which generally corresponds to the time of randomization.}

\item{ctrl_estimates}{Estimated mean outcome in the control group at the
fixed occasions corresponding to the times in \code{time_points}.}

\item{exp_estimates}{Estimated mean outcomes in the experimental group at
fixed occasions corresponding to the times in \code{time_points[-1]}. Note that
first element in \code{exp_estimates} should correspond to some time after
randomization.}

\item{ref_fun}{Intra- and extrapolation function that is returned by
\code{ref_fun_constructor()}.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation.
\item \code{"spline"}: natural cubic spline interpolation. This interpolation method has been most
thoroughly tested is most stable.
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson.
}}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}.}

\item{j}{(Integer) vector that indicates which elements in \code{exp_estimates}
should be used for the score test. Defaults to \code{1:length(exp_estimates)},
i.e., all elements are used.}

\item{weights}{Staring values for the weights vector.}

\item{epsilon}{Tolerance to stop optimization algorithm. If the l2-norm of
the difference in weights from two subsequent iterations is smaller than
\code{epsilon}, the optimization algorithm stops.}
}
\value{
(numeric) vector with optimal weights.
}
\description{
The \code{\link[=optimize_weights]{optimize_weights()}} function finds the weights for the score estimator
that minimize the estimated variance of this estimator. The estimated
variance is based in the delta method.
}
