% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric-bootstrap.R
\name{pm_bootstrap_vertical_to_common}
\alias{pm_bootstrap_vertical_to_common}
\title{Parametric Bootstrap for a Common Acceleration Factor}
\usage{
pm_bootstrap_vertical_to_common(
  TCT_Fit,
  inference,
  B = 100,
  return_se = TRUE,
  null = FALSE,
  select_coef,
  constraints = FALSE,
  weights,
  start_gamma
)
}
\arguments{
\item{TCT_Fit}{Object returned by \code{\link[=TCT_meta]{TCT_meta()}}}

\item{inference}{Which approach is used for estimation and inference? Should
be \code{"least-squares"}, \code{"contrast"}or \code{"delta-method"}.}

\item{B}{Number of bootstrap replications.}

\item{return_se}{(boolean) Return the estimated standard error from each
bootstrap replication? This standard error is computed with the delta method.}

\item{null}{(boolean): conduct the bootstrap under the null hypothesis of no
treatment effect? Defaults to \code{FALSE}.}

\item{select_coef}{Estimates from the \code{exp_estimates} in \code{\link[=TCT_meta]{TCT_meta()}} to use
in estimating the common acceleration factor. If there is reason to believe
that the proportional slowing assumption does not hold for only some time
points (e.g., for the first measurement after randomization), then the
corresponding estimate should not be used in estimation the common
acceleration factor.}

\item{constraints}{Use the constrained generalized least squares estimator
for the vertical treatment effects.}

\item{weights}{If \code{inference = "contrast"}, then the user can provide weights
for the different measurement occasions. If not provided, the optimal
weights are estimated data-adapatively.}

\item{start_gamma}{Initial value for the algorithm that finds the common
acceleration factor numerically. Defaults to \code{0.75}.}
}
\value{
A list with two element:
\itemize{
\item \code{estimates_bootstrap}: (numeric) vector of length \code{B} that contains the
bootstrap replicates for the common acceleration factor.
\item \code{se_bootstrap}: (numeric) vector of length \code{B} that contains the bootstrap
replicates of the estimated standard error. Is a vector of \code{NA}s when
\code{return_se = FALSE}.
}
}
\description{
The \code{\link[=pm_bootstrap_vertical_to_common]{pm_bootstrap_vertical_to_common()}} function implements a parametric
bootstrap for the time-specific acceleration factors.
}
