% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{pm_bootstrap_vertical_to_common}
\alias{pm_bootstrap_vertical_to_common}
\title{Parametric Bootstrap for a Common Acceleration Factor}
\usage{
pm_bootstrap_vertical_to_common(
  time_points,
  ctrl_estimates,
  exp_estimates,
  vcov,
  TCT_vcov,
  interpolation = "spline",
  B = 100,
  bs_fix_vcov = TRUE,
  return_se = TRUE,
  null = FALSE,
  gls_est = TRUE,
  select_coef,
  constraints = FALSE
)
}
\arguments{
\item{return_se}{}
}
\description{
Parametric Bootstrap for a Common Acceleration Factor
}
