% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{pm_bootstrap_vertical_to_horizontal}
\alias{pm_bootstrap_vertical_to_horizontal}
\title{Parametric Bootstrap for Time-Specific Acceleration Factors}
\usage{
pm_bootstrap_vertical_to_horizontal(
  time_points,
  ctrl_estimates,
  exp_estimates,
  vcov,
  interpolation = "spline",
  B = 100
)
}
\arguments{
\item{time_points}{Time points to which the estimates in \code{ctrl_estimates}
correspond.}

\item{ctrl_estimates}{Vector with estimated values for reference trajectory
at \code{time_points}. This vector should thus have the same length as
\code{time_points}.}

\item{exp_estimates}{Vector with estimated values for the active treatment
group at \code{time_points}. This vector should thus have length
\code{length(time_points) - 1} since the baseline values are assumed to be equal
in both treatment groups.}

\item{vcov}{The (estimated) variance-covariance matrix of the parameter
estimates in the \code{c(ctrl_estimates, exp_estimates)} vector.}

\item{interpolation}{Interpolation method; see \code{\link[=TCT]{TCT()}}.}

\item{B}{number of bootstrap replications.}
}
\value{
Matrix where each row corresponds to the bootstrap replicates of the
acceleration factors.
}
\description{
Parametric Bootstrap for Time-Specific Acceleration Factors
}
