% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{print.TCT_meta}
\alias{print.TCT_meta}
\title{Print Meta-Time Component Test object}
\usage{
\method{print}{TCT_meta}(x)
}
\arguments{
\item{x}{}
}
\description{
Print Meta-Time Component Test object
}
