% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{print.TCT_meta_common}
\alias{print.TCT_meta_common}
\title{Print TCT_meta_common object}
\usage{
\method{print}{TCT_meta_common}(x)
}
\arguments{
\item{x}{}
}
\description{
Print TCT_meta_common object
}
