% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-test.R
\name{score_conf_int}
\alias{score_conf_int}
\title{Compute Confidence Interval Based on Score Test}
\usage{
score_conf_int(
  time_points,
  ctrl_estimates,
  exp_estimates,
  ref_fun,
  interpolation,
  vcov,
  j,
  alpha = 0.05
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_estimates} vector.}

\item{ctrl_estimates}{Estimated mean outcome in the control group at fixed
occasions.}

\item{exp_estimates}{Estimated mean outcomes in the experimental group at
fixed occasions.}

\item{ref_fun}{Intra- and extrapolation function that is returned by
\code{ref_fun_constructor()}.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation
\item \code{"spline"}: natural cubic spline interpolation
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson
}}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}.}

\item{j}{Measurement occasion to test acceleration factor for.}

\item{alpha}{\code{1 - alpha} represent the two-sided confidence level. Defaults
to \code{0.05}.}
}
\value{
(numeric) vector with two elements. The first element is the lower
confidence limit, the second element is the upper confidence limit.
}
\description{
The \code{\link[=score_conf_int]{score_conf_int()}} function computes the confidence interval for the
score test at measurement \code{j}.
}
