% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-test.R
\name{score_test_common}
\alias{score_test_common}
\title{Title}
\usage{
score_test_common(
  time_points,
  ctrl_estimates,
  exp_estimates,
  ref_fun,
  interpolation,
  vcov,
  gamma_0 = 1,
  type = "omnibus",
  j = 1:length(exp_estimates),
  weights = NULL
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_estimates} vector.}

\item{ctrl_estimates}{Estimated mean outcome in the control group at fixed
occasions.}

\item{exp_estimates}{Estimated mean outcomes in the experimental group at
fixed occasions.}

\item{ref_fun}{Intra- and extrapolation function that is returned by
\code{ref_fun_constructor()}.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation
\item \code{"spline"}: natural cubic spline interpolation
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson
}}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}.}

\item{gamma_0}{Value under the null hypothesis for the acceleration factor.}

\item{type}{Which type of test statistic should be used. See Details.}

\item{j}{description}

\item{weights}{description}
}
\description{
Title
}
\details{
Four types of test statistics are implemented in the \code{\link[=score_test_common]{score_test_common()}}
function:
\enumerate{
\item \code{type = "omnibus"}: This corresponds to a classic chi-squared test.
\item \code{type = "directional"}:
\item \code{type = "inverse variance"}:
\item \code{type = "custom"}:
}
}
