% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCT-package.R
\docType{package}
\name{TCT-package}
\alias{TCT}
\alias{TCT-package}
\title{TCT: Time Component Tests for Longitudinal Data}
\description{
The goal the TCT package is to provide a set of easy to use functions to perform time components test.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/florianstijven/TCT}
  \item Report bugs at \url{https://github.com/florianstijven/TCT/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
