% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{TCT}
\alias{TCT}
\title{Transform vertical treatment effect estimates to time scale}
\usage{
TCT(
  time_points,
  ctrl_estimates,
  exp_estimates,
  vcov,
  interpolation = "spline",
  B = 0
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_means} vector.}

\item{ctrl_estimates}{Estimated mean outcome in the control group at fixed
occasions.}

\item{exp_estimates}{Estimated mean outcomes in the experimental group at fixed
occasions.}

\item{vcov}{The variance-covariance matrix for the means. In order to map to
the correct estimates, this matrix should be the variance-covariance matrix
of \code{c(ctrl_means, exp_means)}.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation
\item \code{"spline"}: natural cubic spline interpolation
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson
}}

\item{B}{Number of bootstrap replications. If \code{B = 0}, no boostrap is
performed (default).}
}
\value{
An object from the TCT-class:
}
\description{
This function transform the vertical parameter estimates to parameter
estimates on the time scale.
}
