% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_effects.R
\name{TCT_common}
\alias{TCT_common}
\title{Title}
\usage{
TCT_common(
  TCT_Fit,
  B = 0,
  bs_fix_vcov = FALSE,
  select_coef = 1:length(coef(TCT_Fit)),
  null_bs = FALSE,
  gls_est = TRUE
)
}
\arguments{
\item{bs_fix_vcov}{}
}
\description{
Title
}
