% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-test-utils.R
\name{v_deriv_alpha_beta}
\alias{v_deriv_alpha_beta}
\title{Derivative of v-function with respect to \eqn{(\boldsymbol{\alpha}, \boldsymbol{\beta})'}}
\usage{
v_deriv_alpha_beta(
  time_points,
  ctrl_estimates,
  interpolation,
  gamma_0,
  j,
  weights
)
}
\arguments{
\item{time_points}{Ordered vector that contains the times corresponding to
the estimated means in the \code{ctrl_estimates} vector. The first element
should be zero which generally corresponds to the time of randomization.}

\item{ctrl_estimates}{Estimated mean outcome in the control group at the
fixed occasions corresponding to the times in \code{time_points}.}

\item{interpolation}{Which interpolation method to use?
\itemize{
\item \code{"linear"}: linear interpolation.
\item \code{"spline"}: natural cubic spline interpolation. This interpolation method has been most
thoroughly tested is most stable.
\item \verb{"monoH.FC}: monotone Hermite spline according to the method of Fritsch
and Carlson.
}}

\item{gamma_0}{Value for the acceleration factor under the null hypothesis.}

\item{j}{(Integer) vector that indicates which elements in \code{exp_estimates}
should be used for the score test. Defaults to \code{1:length(exp_estimates)},
i.e., all elements are used.}

\item{weights}{If \code{type == "custom"}, the user should specify a weight
vector for weighting estimates at different time points differently.}
}
\value{
(numeric) The derivative described above, evaluated in \code{gamma_0}.
}
\description{
This function computes
\eqn{\frac{\partial v(\boldsymbol{\alpha}, \boldsymbol{\beta}; \gamma_0)}{\partial (\boldsymbol{\alpha}, \boldsymbol{\beta})}}.
}
