% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-test-utils.R
\name{v_deriv_alpha_beta}
\alias{v_deriv_alpha_beta}
\title{Derivative of v-function with respect to \eqn{(\boldsymbol{\alpha}, \boldsymbol{\beta})'}}
\usage{
v_deriv_alpha_beta(
  time_points,
  ctrl_estimates,
  interpolation,
  gamma_0,
  j,
  weights
)
}
\arguments{
\item{gamma_0}{Value for the acceleration factor under the null hypothesis.}

\item{j}{(Integer) vector that indicates which elements in \code{exp_estimates}
should be used for the score test. Defaults to \code{1:length(exp_estimates)},
i.e., all elements are used.}

\item{weights}{If \code{type == "custom"}, the user should specify a weight
vector for weighting estimates at different time points differently.}
}
\description{
This function computes
\eqn{\frac{\partial v(\boldsymbol{\alpha}, \boldsymbol{\beta}; \gamma_0)}{\partial (\boldsymbol{\alpha}, \boldsymbol{\beta})}}.
}
